requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asgiref=3.5.0=pyhd8ed1ab_0
autopep8=1.6.0=pyhd8ed1ab_1
bzip2=1.0.8=h7f98852_4
ca-certificates=2021.10.8=ha878542_0
django=4.0=pyhd8ed1ab_1
django-cors-headers=3.11.0=pyhd8ed1ab_0
djangorestframework=3.13.1=pyhd8ed1ab_0
freetype=2.10.4=h0708190_1
giflib=5.2.1=h36c2ea0_2
jbig=2.1=h7f98852_2003
jpeg=9e=h166bdaf_1
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libdeflate=1.10=h7f98852_0
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_16
libgomp=11.2.0=h1d223b6_16
libnsl=2.0.0=h7f98852_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=he4da1e4_16
libtiff=4.3.0=h542a066_3
libuuid=2.32.1=h7f98852_1000
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libzlib=1.2.11=h166bdaf_1014
lz4-c=1.9.3=h9c3ff4c_1
ncurses=6.3=h27087fc_1
openjpeg=2.4.0=hb52868f_1
openssl=3.0.2=h166bdaf_1
pillow=9.1.0=py310he619898_2
pip=22.0.4=pyhd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
pycodestyle=2.8.0=pyhd8ed1ab_0
python=3.10.4=h2660328_0_cpython
python_abi=3.10=2_cp310
pytz=2022.1=pyhd8ed1ab_0
readline=8.1=h46c0cb4_0
setuptools=62.1.0=py310hff52083_0
sqlite=3.38.3=h4ff8645_0
sqlparse=0.4.2=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
toml=0.10.2=pyhd8ed1ab_0
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=h191b570_0
wheel=0.37.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
zlib=1.2.11=h166bdaf_1014
zstd=1.5.2=ha95c52a_0
